APIv2: Remove "{}" from UUID
The braces don't need to be sent as part of the UUID; in fact, the
Python instagram library [1] does not send them. Moreover, there was a
bug (which this commit fixes) that would remove the braces from the UUID
after posting a photo: the replace() functions modify the string on
which they are invoked.

[1]: https://github.com/LevPasha/Instagram-API-python